Fix field name filter in sidebar. (Graylog2/graylog-plugin-enterprise#389)
Since the introduction of Graylog2/graylog-plugin-enterprise#366, a changing `filter` field in the state
of the `FieldList` component did not lead to a rerender, as the field
was not checked in the `shouldComponentUpdate` function. This lead to
the filter input not having any effect when submitted.

This change is now checking the whole state of the component for
equality, making the filter input usable again.