Add template field validation
The API endpoint requires the template to be present, but until now that
was not checked in the frontend, resulting in a confusing error when
there was no configuration template.

This ensures the field is set before sending the request to the server,
and also sets an error state in that input, to help the user figure out
what happened.