Fix fetching of collector default templates
Before this change, we used the collector list response and tried to get
the default templates for a collector from there. Unfortunately, that
response returns a collector summary, which does not include the default
template, as it may be a large string.

This commit fixes that situation by fetching the selected collector as
soon as we know which collector is selected, and memoizing its default
template so we don't fetch the same collector again.