Isolate search types in separate backend queries. (Graylog2/graylog-plugin-enterprise#392)
Until now, when a single search type failed, the whole query it belongs to failed. The error message returned was not resolvable to the search type which caused it.

With this change, the individual search types of a query generate separate ES queries derived from the ES query generated from the `Query` object the search types belong to. All ES queries are then resolved using a single multi search request. This makes it possible to resolve a failed ES query to the search type that produced it and return its search type id along with a meaningful error message.