Collector Configuration form improvements (#5018)
* Add template field validation

The API endpoint requires the template to be present, but until now that
was not checked in the frontend, resulting in a confusing error when
there was no configuration template.

This ensures the field is set before sending the request to the server,
and also sets an error state in that input, to help the user figure out
what happened.

* Ask user before clearing current configuration

When the user changed the Collector assigned to a Configuration, we
automatically deleted the collector configuration, and tried to set it
to the default template for the Collector (tried but failed, because a bug).

As users may want to keep the current configuration and edit it in the
newly assigned Collector, we now ask users if they want to load the
default template if the configuration input is not empty.

* Fix fetching of collector default templates

Before this change, we used the collector list response and tried to get
the default templates for a collector from there. Unfortunately, that
response returns a collector summary, which does not include the default
template, as it may be a large string.

This commit fixes that situation by fetching the selected collector as
soon as we know which collector is selected, and memoizing its default
template so we don't fetch the same collector again.