Work around Configuration(s)Store/Actions naming inconsistency.
This change is working around a naming inconsistency of the
ConfigurationActions (without s) and ConfigurationsStore (with s) pair.
Before this change, doing

```
  const { ConfigurationActions } = CombinedProvider.get('Configuration')
```

did not load the `ConfigurationsStore`, you had to load it explicitly.

This change makes it possible to load both using `Configuration` or
`Configurations` key passed to `CombinedProvider.get`.