Remove state from TimeUnitInput
`TimeUnitInput` stored state of the input, but it was also relying on
props to get the state right. As all consumers of the component must
provide an `update` function, we can asume they will use that to update
their own state and pass down props with the right values for this
input. That makes it easier to understand and change the input from the
outside, e.g. when props change.