Disable views functionality when no valid license is present. (Graylog2/graylog-plugin-enterprise#419)
This change implements checking for a valid enterprise/views license
before rendering the views management/extended search page. When no
license is found, the license is invalid or an error occurred while
retrieving the licensing information, the functionality is disabled and
instead an alert is shown indicating the reason and (if present) the
error that happened.