Display right collector status on null status code
In some situations a collector's status code may be temporarily `null`.
When we converted that status code into a string, we did an intermediate
conversion into number, which converted `null` into `0`, the status code
for the `running` state. This is not correct, as the status for that
case should be `unknown`.

As I couldn't find any case where the status is a string, this commit
drops the conversion into a number, fixing the issue described above.