Validate collector configuration before saving
Before saving a configuration, render a preview of the template and
throw an exception if there are any errors. This will let the user know
about the situation before some collectors are broken.

This validation will not catch all errors, but at least invalid references
and lexical errors will be found.