Separate stores into own chunk, loaded before plugin/app bundles.
Before this change, any (explicit or implicit) consumer of the
`CombinedProvider` module inevitably included all actions/stores in
core.

With this change, the providers are separated from the inclusion of
actions/stores. Actions/stores are now in a separate 'builtins' bundle
which also registers these in the `CombinedProvider` module upon entry.