Randomize chunk id only once.
Before this change, the chunk id of a webpack compilation was being
prefixed with a random, build-specific identifier. This happened for
every compilation, even partial ones (when changing a file during
webpack/webpack-dev-server watch mode). Unfortunately there was no check
if the prefix had already been added, so for every partial compilation
it was added all over again.

This PR introduces a check if the given chunk contains an id which is
already prefixed and skips prefixing it in this case.