Fix content pack serialisation
A downloaded content pack had *AutoValue_ContentPackV1* as version
serialized in the json. That lead to errors when the content pack
was downloaded and later uploaded again.

The problem was that `JsonTypeInfo.as` was using `PROPERTY`
which would read the version to determine the version of an
content pack, but also write the class name into the version.

https://fasterxml.github.io/jackson-annotations/javadoc/2.3.0/com/fasterxml/jackson/annotation/JsonTypeInfo.As.html

We now use `EXISTING_PROPERTY` so `JsonTypeInfo.as` will only
read this value but not write it itself.