Add more lifecycle indicators to PageHeader
`PageHeader` so far allowed us to set an `experimental` flag, that
would show an indicator if the rendered page was still considered
experimental.

This change adds a `legacy` lifecycle as well, and makes setting
lifecycles more generic, so we could add more in the future if needed.

This commit also adds the possibility to set custom lifecycle messages
that will be displayed in the tooltip, so consumers of `PageHeader` can
customize the default messages for the lifecycle set.