Add title to NativeEntityDescriptor
The NativeEntityDescriptor holds all the information
needed to display a installed entity. Only the title is missing.

Instead of creating a new object which would be filled on the fly
we save the title of the entity next to id and type.

This should keep the complexity low and calculation time
to find a name as well