Remove warnings from test (#5142)
* Add required fields to EditPatternModal.test

The test was missing from the component required props.
Instead of providing a default I add stubs to the test
and mark testPattern as required as well. Since beside the test
this three functions a vital for the Modal

* Use string format instead of numbers

The WidgetFooter.test was using utc date numer to display
in the footer. But the footer was requireing a string.
Use ISOString format instead

* Add required props to DelegatedSearchPage

The DelegatedSearchPage.test was missing two props
if called without a component, since the underlying SearchPage
needed them

* Add revision to content pack

The ContentPackDownloadControls require a revision so
the ContentPackList.test needed a rev in his content packs

* Wrap MessageTableEntry in <table/>

Enzymes 'mount()' will wrap the given component into
a <div /> which will raise a warning. Therefore
I wrap MessageTableEntry in a <table />

* use mockComponent to mock ClipboardButton

The ClipboardButton mocking was raising an warning.
To eliminate the warning I renamed the ClipboardButton
to clipboard-button. As seen in:

https://stackoverflow.com/questions/50253660/mock-with-jest-causes-warnings-correct-casing-tag-is-unrecognized-and-unknown

* Fail when console.error ist called

* Mark uncheckable ExpandlebListItem as readOnly

A ExpandleListItem which is not selectable
should be marked a readOnly so it does not raise a warning

* Prefill content pack to avoid warning

If the input fields are not prefilled with empty strings,
react will warn that the states of the inputs are changed from
uncontrolled to controlled input.

* Mock console.error

we want to test that console.error was called rightfully,
but don't want console.error to be seen on the test output.

* Fix error formating