Warn about unsaved changes in views when leaving route. (Graylog2/graylog-plugin-enterprise#431)
* Warn about unsaved changes in views when leaving route.

Before this change, a hook was registered (but not properly
unregistered) on window when the current page was left (by reloading the
page or going back/forth in browser history). This hook was not called,
when the current route in the application was left (e.g. by clicking a
an item in the navigation bar).

This change is now registering a hook with react-router that is called
when the current route is left, which does the exact same page and shows
a confirmation dialog warning about unsaved changes and offers the
option to stay on the current page by clicking cancel.

* Adding tests for OnSaveViewAction.

* Using bluebird promises in tests.

* Do not return promise if not present.

* Adding test for ExtendedSearchPage.

* Adding tests for WindowLeaveMessage.