Fixing Jest for Babel 7 in Plugins.
Due to a change in Babel 7 that affects resolution of babel configs[1]
(`.babelrc` is now a hierarchical config, while `babel.config.js` is
not), source files outside of the source root of jest will not be
transpiled. That means, plugins running tests were not able to
require/import source files of core.

With this change, `.babelrc` will be renamed to `babel.config.js`, so it
is still used when a plugin imports a file from core.

[1]: https://github.com/facebook/jest/issues/6053#issuecomment-383632515