Implement Error Boundary for Navigation Bar children.
This change implements a error boundary component wrapping children of
the navigation bar in the `App` component.

Before this change, any errors in children components of the navigation
bar caused the application to malfunction, errors being shown on the
js console only and (for React 16) the DOM to be unmounted, requiring
the user to do a full page refresh to be able to work with the
application again.

After this change, errors occuring in components which are children of
the navigation bar leave the application in a functional state, while
showing the error message, stack trace and component tree in the
browser. Users are still able to navigate the application afterwards.