Improving MockComponent by rendering custom component. (#5161)
This change is improving the `MockComponent` function which generates a
mock component usable in tests when rendering a subtree is not desired.

Before this change, `MockComponent` generated a component which renders
a series of `span` elements which were carrying the passed `props` to
it, so they show up in a snapshot. Unfortunately this fails when React
warnings fail tests, because those props were not accepted on `span`
elements.

After this change, `MockComponent` returns a component rendering a
custom element, which is named after the string parameter passed to it,
e.g. `MockComponent('FooMock')` returns a component rendering

```
<foo-mock {...props}>
  {children>
</foo-mock>
```

which serves the purpose of creating a minimal element which carries all
props and children passed to it, so they show up in a snapshot, in
contrast to a common pattern of using `() => 'foo-mock'` instead.