Catch Netty exceptions and avoid logging stacktraces
To handle exceptions that happen in the server context,
we need to register our ExceptionLoggingChannelHandler as a childHandler.

In addition, make sure that the ExceptionLoggingChannelHandler is
always added to the top of the (bottom-up evaluated) pipeline.
Otherwise it might miss some exceptions.

Do not propagate the exception any further.
A full netty stacktrace does not help anyone.
One meaningful log message should be enough.
This last bit of the change fixes #5109