Fixes for content packs (#5153)
* Allign Install form fields

The install form has a list of entities
which were unallgined to the form fields.

Fixes #5115

* Fix closing of ContentPackEditParameter modal

the modal did not close since it was a child of
the open button. A click on the close button bubbled up
to the open button which would than reopen the closed
modal.

I moved out the modal as a child from the open button.

* Make click on ExpandleListItem select checkbox

We do not have label and checkboxes, since we do not have
ids for the ExpandleListItem. There for I add an click
handler to the header which will click on the
associated checkbox.

Fixes #5114

* Fix annotation from @edmundoa

Move click handler to its own named function to prevent
function createion leaking