Return helpful error message when CSRF protection fails.
Before this change, the server returned a generic "Bad Request" error
message when a client queried the API without the required CSRF
protection header. This leaves the consumer confused, especially for
previous users of the API which now do not know what to change in their
request to make it work again.

This change now leaves the error code intact but adds a helpful message
telling the consumer what went wrong.

Fixes #5012.