Catch errors when no type information for field is passed to `Value`. (Graylog2/graylog-plugin-enterprise#447)
Before this change, any render of the `Value` component without a type
information passed for the field and no `children` resulted in an error.
This could happen when rendering the data table header columns in
situations where the configuration of an aggregation was recently
changed but no updated result was present (yet).

This change now sets a default prop of `FieldType.Unknown` for the
`type` prop of the component to prevent accessing a `type` field for
`undefined`. Additionally it fixes rendering the value without a
type-specific component in the dropdown and adds tests.

Fixes Graylog2/graylog-plugin-enterprise#446.