Add graylog fields to beats default templates
The collector "Show messages" button links to a search
for a matching `gl2_source_collector`.
Thus we need to define these fields in every beats configuration.

Furthermore, we need to set `fields_under_root` because the
new Beats input does not strip away the "fields_" prefix for us.
Only "Beats Legacy" does that.