Fixing editing outputs from web and REST API. (#5163)
Before this change, due to a mistake in a former PR (#4753), updating an
output was using the wrong field to find the corresponding document to
be modified. This lead to a successfull request, but no changes were
done in the database.

This was fixed and the old behavior was restored. Additionally, instead
of returning the old object from the database and back to the API, the
new, modified one is returned and upserting is explicitly disabled.