Introduce alert condition subscriptions for alarm callbacks
Alarm callback configurations get a new field that defines which alert
conditions the callback is subscribed to. This can be used to decide, if
a notification should be triggered or not.

Existing alarm callbacks that don't have the new subscription field yet
will be subscribed to all alert conditions to keep the current behavior.

This only implements the backend part and doesn't change any UI yet.