Prevent overwriting children's props in `IfPermitted`.
`IfPermitted` passes given props to children to allow it being used in
scopes where the wrapped child(ren) is/are supposed to follow directly
after a parent component.

Unfortunately this mechanic previously did not check if the passed
property is already present on the child. This lead to a regression
where event handlers where overwritten.

This change is now checking if the given prop is present on the child
before passing it.