Show empty result component in static message list only. (Graylog2/graylog-plugin-enterprise#462)
* Show empty result component in static message list only.

Before this change, the empty result component was shown in every widget
that had no data available. In most of the cases this was the result of
the current query returning 0 documents, so the component would be shown
(and therefore duplicated) in every widget for this query.

After this change, the empty result component is shown only in the
static message list. All other widgets are showing empty graphs/0 (in
the case of the single number widget) instead.

* Adapting tests.