Avoid using `naturalSort` for sorting field names. (Graylog2/graylog-plugin-enterprise#469)
* Avoid using `naturalSort` for sorting field names.

Before this change, `naturalSort` was used in several places to sort
field names or pivots. As the number of field names is generally
unbounded and depends on the environment the product is run in, it can
easily be in the order of several thousands. Therefore, using
`naturalSort` can reduce in a performance impact due to its excessive
use of regular expressions.

This change is removing the usage of `naturalSort` for these cases and
uses a generic JS implementation. This lacks a lot of the features of
`naturalSort`, but the most important one (numeric comparison) can be
used by supplying `{ numeric: true }` as options hash for the generator
function. It is currently not used, as it is not considered to be
important for field names. Also, sorting field names is now
case-insensitive, as it is less surprising for the user.

Fixes Graylog2/graylog-plugin-enterprise#468.

* Explaining the usage of the `DefaultCompare` module.