Allow canceling edit mode abandoning changes. (Graylog2/graylog-plugin-enterprise#487)
* Adding option to cancel editing a widget.

* Allow canceling edit mode abandoning changes.

When editing a widget, the user might to decide to abandon the changes
made and return to the original state of the widget. Before this change
this was not possible, as every change done in edit mode was changing
the underlying data structures.

With this change, starting edit mode keeps a copy of the widget's
config. When the user decides to cancel edit mode instead of finishing
it, the old widget config will be restored.

Fixes Graylog2/graylog-plugin-enterprise#340.