Make search query in alert conditions configurable in the UI (#5212)
* Make search query in alert conditions configurable in the UI

This exposes a new query config field in the alert condition
configuration. The value defaults to "*".

The query is used to build a filter string for the alert condition
search. We add the query string to the query filter for performance
reasons.

Fixes #3966

* Trim streamId and query strings

* Use query config in field content value condition as well

* Add HTTP API endpoints to test alert conditions

There are two test endpoints, one to test existing alert conditions and
one to test new alert conditions which have not been stored in the
database yet.

* Add missing NoAuditEvent annotations to test resources

* Fix failing test because of missing stream id

* Use AND operator and parentheses to combine streams and custom query

Otherwise the generated query will be wrong.

(cherry picked from commit 0533cfd57b2efe2425563074d7e47ffa8468d25c)