Change title of "Add Widget" button to "Create". (Graylog2/graylog-plugin-enterprise#501)
* Change title of "Add Widget" button to "Create".

To prevent limiting the function of the Dropdown menu used for creating
new entities to widgets, the title of it is changed from "Add Widget" to
"Create". This allows us to place functionality in there which creates
non-widget, e.g. a "Parameter" option that allows users to declare
parameters without having to use them first and therefore makes the
availability of parameters visible to the user.

Additionally, the dropdown list is separated into what are "presets" in
the upper half of the list (including a message count and a message
table, more to come) and generic entities (namely "Custom Aggregation")
in the lower half of it.

* Adapting tests.