Allow showing/hiding BootstrapModalForm declaratively.
Before this change, the `BootstrapModalForm` component relied on the
consumer calling the `open`/`close` functions, which does not align well
with the way react is used declaratively.

This change leaves this functionality in place but also adds a `show`
property (being `false` by default) which allows to show a modal when
the property is `true`.