Allow passing container prop to `ClipboardButton`. (#5301)
* Updating clipboard.js to 2.x.

This change is updating clipboard.js to 2.x. The only change is the
renaming of the class from `Clipboard` to `ClipboardJS` (which is not
even relevant for us due to using the ES6 default import), which is
reflected in this PR for consistency.

* Allow passing container prop to `ClipboardButton`.

Under some circumstances (e.g. when embedding a `ClipboardButton` in the
body of a modal dialog), copying a value does not work because it
requires the currently focussed container.

This change allows passing the current container to `ClipboardButton`,
therefore enabling it to work properly in e.g. modal dialogs.

* Adding simple tests.