Fix slow CSV export by using a bigger default batch size (#5289) (#5304)
The default batch size in Elasticsearch is 10. We now use 500 as default
for the newly introduced "batch_size" parameter which we pass to the
search request as limit. (which is the batch size for scroll requests)

Previously we used the "limit" parameter for the batch size AND the
result limit. That was wrong and resulted in truncated results.

Fixes #5172

(cherry picked from commit a77756d54ce27bea15e921e0c75c0f34e19b1a40)