Update variables in a configuration form on a rename
If the user is editing a configuration and changes the names
of variables that he uses, find and replace them directly after
we saved this variable.
We have to keep unsaved changes the user might have done up to this
point, so we cannot simply reload the updated configuration from the
server.