Fix ES alias request for Indices#getIndexNamesAndAliases() method (#5339)
ES 5 and earlier didn't return closed indices when using the "/_alias"
API by default. In ES 6 that behavior changed to return all indices by
default. (even closed ones)

ES 6 effectively changed the default value for the "expand_wildcards"
parameter from "open" to "all" so we have to explicitly set it to "open"
to restore the old behavior.

While we are here, fix an old bug where the set of alias names for an
index was set to the string "alias" instead of the actual alias names.
This was probably introduced when we switched to the ES HTTP API but
never had any negative effect because we use the alias names only in
index retention and there it had no negative effect unless you set the
max number of indices to 0. (which no one does)

Also leave a TODO to improve the efficiency of the Indices#aliasTarget()
method.

Fixes #5336

(cherry picked from commit c2e0410335b933b49b11658df484e9932f32fbda)