Guarding against `undefined` in `isPermitted`/`isAnyPermitted`. (#5328)
Before this change, passing `undefined` for the permission set that
`isPermitted`/`isAnyPermitted` check against, both functions would throw
an exception.

After this change, both functions return proper results based on the set
of required permissions. If it is empty, the functions return `true`, if
it is non-empty, both return `false`.

Fixes #5193.