Add scaling factor for auto interval ranges, doubling average buckets. (Graylog2/graylog-plugin-enterprise#529)
* Add scaling factor for auto interval ranges, doubling average buckets.

This change is doing two things:

1.) Introducing an internal scaling factor. The duration of any given
time range is multiplied with it, before it is matched against the set
of ranges. This allows to fine-tune the average number of buckets across
the range of durations. While it is not exposed currently, it is planned
to make it user-configurable in the web UI, so users can scale the
number of buckets returned for a given timestamp pivot while it is still
scaled when the timerange is changed.

2.) The internal default scaling factor is changed from 1.0 to 0.5,
leading to an effective doubling of returned buckets, increasing
resolution for plots.

* Ensuring `scaling` is set.

* Adapting tests to new output of toString().