Remove nonexistent "consistency" parameter from index/bulk requests
The parameter has been removed from Elasticsearch about 2.5 years ago
before the release of 5.0 and requests have have it will fail.

This hasn't been noticed so far because it only happens when the
Indices#move() method runs. That method is only called by the
FixDeflectorByMoveJob periodical. The job doesn't run that often,
apparently.

Also the only test that is using it is the MessagesIT test and that is
using "assumeTrue()" so it doesn't fail but gets skipped.