Decoupling search creation and search execution. (Graylog2/graylog-plugin-enterprise#538)
* Adding types.

* Decoupling search creation and search execution.

This change decouples search creation and search execution. Before this
change, a search was created and executed in one go. This lead to
problems when a parameter without a value was declared, because the
search could not be saved as it was not executable at this point. This
lead to the weird scenario of having to declare parameter, insert a
value, execute a search and then it would be possible to save the view.
If the view was saved earlier, it would have been corrupted due to a
search being referenced that does not exist.

This search is now creating the search whenever a structural change to
the internal search object happens. If the search result is refreshed,
it is executed.

* Adding bluebird types for flow.

* Adapting tests to guard clause in `SearchStore.execute`.

* Streamline parameter/search execution state change propagation.

This change is destructuring the component tree and limiting rerendering
due to changed search execution state to their smallest scope instead of
rerendering the whole search page. This reduces input lag immensely.

* Minimizing functions, removing linter hint.

* Do not rely on value propagated back in props.

The `value` of the `QueryInput` component is kept in its `state`,
additionally to being updated periodically through `props`. This
increases input lag and leads to situations where `this.state.value` is
more recent than `this.props.value` and part of the user's input is
removed again.

This change is removing updating `state.value` through `props.value`,
assuming that it is always up to date. Additionally, it expects
`props.onChange` to return a promise, which is then used to sync
executing `onExecute` after its completion.

* Saving connect-ed components into separate files, adapt tests.

* Adding tests for ViewStore.

* Adapt result histogram creation to builders.

* Adding more ViewStore test cases.

* Prevent jumping cursor in parameter inputs.