Add Test
This Change adds tests for grok pattern dependecy resolution.

To use the grok pattern extractor in the inputs.json
this change changes the conversion from number to long.
Since the mongodb driver will return the order as long
only a cast was necessary. But now I use a sample
db in json. And json does not support long. So I cast
the order value (which is only -1, 0, 1) to Number and
than convert it to long.