Add ValueReferenceData class and test
This can take an entity data object and can generate config paths as
well as modify data for a path.

It encapsulates all knowledge about the ValueReference objects from the
server to avoid duplicating this in the rest of the code base.

This is a replacement for ObjectUtils#getPaths, #getValue and #setValue.