Remove constraints from Content Pack; add a getter
To move the constraints from top level content pack
to the matching entity this change removes
the requires field from the content pack and
adds a getter method wich collects all constraints
from all entities. That way we can check the
constraints of a content pack easily.

This PR also removes the EntityWithConstraints
from the Facads.

Fixes: #5335