Remove nonexistent "consistency" parameter from index/bulk requests (#5389)
* Remove nonexistent "consistency" parameter from index/bulk requests

The parameter has been removed from Elasticsearch about 2.5 years ago
before the release of 5.0 and requests have have it will fail.

This hasn't been noticed so far because it only happens when the
Indices#move() method runs. That method is only called by the
FixDeflectorByMoveJob periodical. The job doesn't run that often,
apparently.

Also the only test that is using it is the MessagesIT test and that is
using "assumeTrue()" so it doesn't fail but gets skipped.

* Change MessagesIT to actually fail when the request didn't succeed

It was using "assumeTrue()" before so the test got marked as skipped
instead of failed if the request failed.