Persist viewport of Map Visualization when in edit mode. (Graylog2/graylog-plugin-enterprise#563)
* Adding object allowing to define and persist visualization configuration

* Adding visualization config for world map.

* Tracking viewport in component.

* Adding structured classes for viewport of map.

* Correcting flow types for subtyping.

* Using Viewport in map visualization config.

* Using viewport in JS class for map visualization, allowing subtyping.

* Adding visualization config to aggregation widget config.

* Registering world map config subtype.

* Removing unnecessary `toObject` call.

* Correcting field name, checking for null.

* Restoring original order of constructor parameters.

* Adding JSON field names to Viewport properties.

* Implementing equals for AggregationWidget.

* Correcting field name for JSON deserialization of AggregationWidgetConfig

* Extending VisualizationConfigClass, implementing getter for viewport.

* Adding type for getInitialState in ViewStore.

* Registering VisualizationConfig subtype in bindings.

* Adding callbacks for config changes to visualization.

This change adds props to the world map visualization components which
allow passing back changes to the visualization config. The
visualization config is different from the widget config, it includes
configuration overriding parts of the appearance of the selected
visualization.

* Implement `equals` for AggregationWidgetConfig.

This function compares two `AggregationWidgetConfig` instances,
not taking the visualization type or config into account. This is
because it is meant to be used to decide if _structural_ changes
relevant for search generation happened. This helps us to regenerate the
search object and perform a new search execution only if the search is
different, which is not the case when properties of the presentation
have changed.

* Pass visualization type when deserializing visualization config object.

As the type of the visualization is not included in the visualization
config object, we need to pass it separately.

* Only refresh search when the search of the view was updated.

This prevents unnecessary search executions when presentational
properties (changes of a aggregation builder's visualization type) were
changed, but no structural changes were made to the search.

* Updating/adding test cases for ExtendedSearchPage.

* Updating/adding test case for AggregationControls.

* Adding test cases for AggregationBuilder checking passing of onChange.

* Removing unused import.