Content Pack Dependency Resolution for Grok Extractors in Inputs (#5400)
* Move resolve of lookuptable in its own function

This change prepares the resolveNativeEntity and
resolveForInstallation function to calcluate the
dependencies of inputs by diffrent kinds of entities.

This change moves the lookup table relevante code
in its own function.

* Add function for resolving grok pattern dependencies

A input has grok patterns as dependency via extractors.
If a user wants to add the input into a content pack
he also wants to add the refered grok patterns.

This change adds the dependency resolution to
resolveNativeEntity and resolveForInstallation.

* Add test resource

* Add Test

This Change adds tests for grok pattern dependecy resolution.

To use the grok pattern extractor in the inputs.json
this change changes the conversion from number to long.
Since the mongodb driver will return the order as long
only a cast was necessary. But now I use a sample
db in json. And json does not support long. So I cast
the order value (which is only -1, 0, 1) to Number and
than convert it to long.

* Use call by refernece for mutable graph

Prior to this change, the mutable graph was
returned by resolving functions and given
again as a parameter to the next resolving function.

This change will use call by reverence instead
which is less intuitive to read but avoids a lot
of code clutter.

* Add comment to describe nature of cast

The cast to number and eventually to long is
a bit unsual, so I add a comment to clarify
my motivation.

Refs #5399