Rename ValueTyped and ValueReference fields
Prior to this change, the json fields of ValueTyped and
ValueReference were "value" and "type" which made it hard
to distinquish between normal fields and ValueReference fields.

This change will rename the fields to @value and @type.

To prevent to add @ to the type of a parameter we remove
the `extends ValueTyped` from Parameter and instead
implement getter and setter by hand.