Fix display of title and description
Prior to this change, the changed naming of type and value
prevented the display of title and description in the
entity list.

This change uses the new ValueRefHelper to test if the
value is a ValueRef and extracts the value.

Additionally the change adjust the ValueRefHelper to be
used by a normal object.