Disable content pack parameter type when editing
Prior to this change, the type of a parameter could be changed
which led to problems if the parameter was already applied to
configuration values. Since the value type and the parameter type
must match.

This change will prevent the change of a parameter type. If the user
wishes the change of a parameter type, he has to delete and recreate
a parameter.